% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ALE}
\alias{ALE}
\title{Accumulated Local Effect Plot (ALE) for one feature}
\usage{
ALE(model, variable, neighborhoods = 10)
}
\arguments{
\item{model}{a model created by \code{\link{dnn}}}

\item{variable}{variable as string for which the PDP should be done}
}
\description{
Does the ALE for one feature
}
\examples{
\dontrun{
library(cito)
set.seed(222)
validation_set<- sample(c(1:nrow(datasets::iris)),25)

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris[-validation_set,],validation = 0.1)

# show zoomable plot of training and validation losses
analyze_training(nn.fit)

# set model which is used for predictions to model from epoch 22
nn.fit$use_model_epoch <- 22

# Use model on validation set
predictions <- predict(nn.fit, iris[validation_set,])

# Scatterplot
plot(iris[validation_set,]$Sepal.Length,predictions)
}
}
