% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{analyze_training}
\alias{analyze_training}
\title{Visualize training of Neural Network afterwards to decide on best performing model.
Creates a plotly figure which allows to zoom in and out on training graph}
\usage{
analyze_training(object)
}
\arguments{
\item{object}{a model created by \code{\link{dnn}}}
}
\value{
a plotly figure
}
\description{
Visualize training of Neural Network afterwards to decide on best performing model.
Creates a plotly figure which allows to zoom in and out on training graph
}
