% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{avgPool}
\alias{avgPool}
\title{This function creates an \code{avgPool} layer object of class \code{citolayer} for use in constructing a Convolutional Neural Network (CNN) architecture. The resulting layer object can be passed to the \code{\link{create_architecture}} function to define the structure of the network.}
\usage{
avgPool(kernel_size = NULL, stride = NULL, padding = NULL)
}
\arguments{
\item{kernel_size}{(integer or tuple) The size of the kernel in this layer. Use a tuple if the kernel size differs across dimensions.}

\item{stride}{(integer or tuple) The stride of the kernel in this layer. If \code{NULL}, the stride is set to the kernel size. Use a tuple if the stride differs across dimensions.}

\item{padding}{(integer or tuple) The amount of zero-padding added to the input on both sides. Use a tuple if the padding differs across dimensions.}
}
\value{
An S3 object of class \code{"avgPool" "citolayer"}, representing an average pooling layer in the CNN architecture.
}
\description{
This function creates an \code{avgPool} layer object of class \code{citolayer} for use in constructing a Convolutional Neural Network (CNN) architecture. The resulting layer object can be passed to the \code{\link{create_architecture}} function to define the structure of the network.
}
\details{
This function creates an \code{avgPool} layer object, which represents an average pooling layer in a CNN architecture. Parameters not specified (and thus set to \code{NULL}) will be filled with default values provided to the \code{\link{create_architecture}} function.
}
\examples{
\donttest{
if(torch::torch_is_installed()){
library(cito)

# A average pooling layer where all available parameters are assigned
# No value will be overwritten by 'create_architecture()'
layer1 <- avgPool(3, 1, 0)

# A average pooling layer where only the kernel size is assigned
# stride and padding are filled with the defaults
# passed to the 'create_architecture()' function
layer2 <- avgPool(kernel_size=4)
}
}
}
\seealso{
\code{\link{create_architecture}}
}
\author{
Armin Schenk
}
