% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model}
\alias{check_model}
\title{check if model is loaded and if current parameters are the desired ones}
\usage{
check_model(object)
}
\description{
check if model is loaded and if current parameters are the desired ones
}
