% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{coef.citocnn}
\alias{coef.citocnn}
\title{Returns list of parameters the neural network model currently has in use}
\usage{
\method{coef}{citocnn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{cnn}}}

\item{...}{nothing implemented yet}
}
\value{
list of weights of neural network
}
\description{
Returns list of parameters the neural network model currently has in use
}
