% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{coef.citodnn}
\alias{coef.citodnn}
\title{Returns list of parameters the neural network model currently has in use}
\usage{
\method{coef}{citodnn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{dnn}}}
}
\value{
list of weigths of neural network
}
\description{
Returns list of parameters the neural network model currently has in use
}
