% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmn.R
\name{coef.citommn}
\alias{coef.citommn}
\title{Retrieve parameters of a fitted MMN model}
\usage{
\method{coef}{citommn}(object, ...)
}
\arguments{
\item{object}{A model created by \code{\link{mmn}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list with two components:
\itemize{
\item \code{net_parameters}: A list of the model's weights and biases for the currently used model epoch.
\item \code{net_buffers}: A list of buffers (e.g., running statistics) for the currently used model epoch.
\item \code{loss_parameters}: A list of the loss function's parameters for the currently used model epoch.
}
}
\description{
This function returns the list of parameters (weights and biases) and buffers (e.g. running mean and variance of batch normalization layers) currently in use by the neural network model created using the \code{\link{mmn}} function.
}
