% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_scheduler.R
\name{config_lr_scheduler}
\alias{config_lr_scheduler}
\title{Creation of customized learning rate scheduler objects}
\usage{
config_lr_scheduler(
  type = c("lambda", "multiplicative", "one_cycle", "step"),
  ...
)
}
\arguments{
\item{type}{string defining which type of scheduler should be used. See Details.}

\item{...}{additional arguments to be passed to scheduler. See Details.}
}
\value{
object of class cito_lr_scheduler to give to \code{\link{dnn}}
}
\description{
Helps create custom learning rate schedulers for \code{\link{dnn}}.
}
\details{
different learning rate scheduler need different variables, these functions will tell you which variables can be set:
\itemize{
\item lambda: \code{\link[torch]{lr_lambda}}
\item multiplicative: \code{\link[torch]{lr_multiplicative}}
\item one_cycle: \code{\link[torch]{lr_one_cycle}}
\item step: \code{\link[torch]{lr_step}}
}
}
\examples{
\donttest{
library(cito)

# create optimizer object
opt <- config_optimizer(type = "adagrad", lr_decay = 1e-04, weight_decay = 0.1 )

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris, optimizer = opt)

}
}
