% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_optimizer}
\alias{config_optimizer}
\title{Creation of customized optimizer objects}
\usage{
config_optimizer(
  optim = c("adam", "adadelta", "adagrad", "rmsprop", "rprop", "sgd"),
  ...
)
}
\arguments{
\item{optim}{character string defining the type of optimizer}

\item{...}{additional arguments to be passed to optimizer. See Details.}
}
\value{
object of class cito.optim to give to \code{\link{dnn}}
}
\description{
Helpd you create custom optimizer for \coode{\link{dnn}}. It is recommended to set learning rate in \coode{\link{dnn}}.
}
\details{
different optimizer need different variables, this function will tell you how the variables are set.
For more information see the corresponding functions:
\itemize{
\item adam: \code{\link[torch]{optim_adam}}
\item adadelta: \code{\link[torch]{optim_adadelta}}
\item adagrad: \code{\link[torch]{optim_adagrad}}
\item rmsprop: \code{\link[torch]{optim_rmsprop}}
\item rprop: \code{\link[torch]{optim_rprop}}
\item sgd: \code{\link[torch]{optim_sgd}}
}
}
