% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{e}
\alias{e}
\title{Embeddings}
\usage{
e(dim = 1L, weights = NULL, train = TRUE, lambda = 0, alpha = 1)
}
\arguments{
\item{dim}{integer, embedding dimension}

\item{weights}{matrix, to use custom embedding matrices}

\item{train}{logical, should the embeddings be trained or not}

\item{lambda}{regularization strength on the embeddings}

\item{alpha}{mix between L1 and L2 regularization}
}
\description{
Can be used for categorical variables, a more efficient alternative to one-hot encoding
}
