% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multinomial_log_prob}
\alias{multinomial_log_prob}
\title{Multinomial log likelihood}
\usage{
multinomial_log_prob(probs, value)
}
\arguments{
\item{probs}{probabilities}

\item{value}{observed values

Multinomial log likelihood}
}
\description{
Multinomial log likelihood
}
