% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{plot.citoarchitecture}
\alias{plot.citoarchitecture}
\title{Plot the CNN architecture}
\usage{
\method{plot}{citoarchitecture}(x, input_shape, output_shape = NULL, ...)
}
\arguments{
\item{x}{an object of class citoarchitecture created by \code{\link{create_architecture}}}

\item{input_shape}{a vector with the dimensions of a single sample (e.g. c(3,28,28))}

\item{output_shape}{the number of nodes in the output layer}

\item{...}{additional arguments}
}
\value{
nothing
}
\description{
Plot the CNN architecture
}
