% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{plot.citocnn}
\alias{plot.citocnn}
\title{Creates plot which gives an overview of the network architecture.}
\usage{
\method{plot}{citocnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
nothing
}
\description{
Creates plot which gives an overview of the network architecture.
}
