% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{plot.citodnn}
\alias{plot.citodnn}
\title{Creates graph plot which gives an overview of the network architecture.}
\usage{
\method{plot}{citodnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{dnn}}}

\item{...}{no further functionality implemented yet}
}
\description{
Creates graph plot which gives an overview of the network architecture.
}
\examples{
\dontrun{
library(cito)

set.seed(222)
validation_set<- sample(c(1:nrow(datasets::iris)),25)

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris[-validation_set,])

plot(nn.fit)
}
}
