% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{predict.citodnn}
\alias{predict.citodnn}
\title{Predict from a fitted dnn model}
\usage{
\method{predict}{citodnn}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{a model created by \code{\link{dnn}}}

\item{newdata}{newdata for predictions}

\item{type}{link or response}

\item{...}{additional arguments}
}
\value{
prediction matrix
}
\description{
Predict from a fitted dnn model
}
