% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.avgPool}
\alias{print.avgPool}
\title{Print pooling layer}
\usage{
\method{print}{avgPool}(x, input_shape, ...)
}
\arguments{
\item{x}{an object of class avgPool}

\item{input_shape}{input shape}

\item{...}{further arguments, not supported yet}
}
\description{
Print pooling layer
}
