% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{print.citodnn}
\alias{print.citodnn}
\title{Print class citodnn}
\usage{
\method{print}{citodnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{dnn}}}

\item{...}{additional arguments}
}
\value{
prediction matrix

original object x gets returned
}
\description{
Print class citodnn
}
\examples{
\donttest{
if(torch::torch_is_installed()){
library(cito)

set.seed(222)
validation_set<- sample(c(1:nrow(datasets::iris)),25)

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris[-validation_set,])

# Sturcture of Neural Network
print(nn.fit)
}
}
}
