% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmn.R
\name{print.citommn}
\alias{print.citommn}
\title{Print class citommn}
\usage{
\method{print}{citommn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{mmn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
Print class citommn
}
