% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.conv}
\alias{print.conv}
\title{Print conv layer}
\usage{
\method{print}{conv}(x, input_shape, ...)
}
\arguments{
\item{x}{an object of class conv}

\item{input_shape}{input shape}

\item{...}{further arguments, not supported yet}
}
\description{
Print conv layer
}
