% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.linear}
\alias{print.linear}
\title{Print linear layer}
\usage{
\method{print}{linear}(x, input_shape, ...)
}
\arguments{
\item{x}{an object of class linear}

\item{input_shape}{input shape}

\item{...}{further arguments, not supported yet}
}
\description{
Print linear layer
}
