% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{print.summary.citodnn}
\alias{print.summary.citodnn}
\title{Print method for class summary.citodnn}
\usage{
\method{print}{summary.citodnn}(x, ...)
}
\arguments{
\item{x}{a summary object created by \code{\link{summary.citodnn}}}

\item{...}{additional arguments}
}
\description{
Print method for class summary.citodnn
}
