% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{print.summary.citodnnBootstrap}
\alias{print.summary.citodnnBootstrap}
\title{Print method for class summary.citodnnBootstrap}
\usage{
\method{print}{summary.citodnnBootstrap}(x, ...)
}
\arguments{
\item{x}{a summary object created by \code{\link{summary.citodnnBootstrap}}}

\item{...}{additional arguments}
}
\value{
List with Matrices for importance, average CE, absolute sum of CE, and standard deviation of the CE
}
\description{
Print method for class summary.citodnnBootstrap
}
