% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.transfer}
\alias{print.transfer}
\title{Print transfer model}
\usage{
\method{print}{transfer}(x, input_shape, output_shape, ...)
}
\arguments{
\item{x}{an object of class transfer}

\item{input_shape}{input shape}

\item{output_shape}{output shape}

\item{...}{further arguments, not supported yet}
}
\description{
Print transfer model
}
