% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{residuals.citodnn}
\alias{residuals.citodnn}
\title{Extract Model Residuals}
\usage{
\method{residuals}{citodnn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{dnn}}}

\item{...}{no additional arguments implemented}
}
\value{
residuals of training set
}
\description{
Returns residuals of training set.
}
