% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{summary.citocnn}
\alias{summary.citocnn}
\title{Summarize a fitted CNN model}
\usage{
\method{summary}{citocnn}(object, ...)
}
\arguments{
\item{object}{A model created by \code{\link{cnn}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The original model object \code{object}, returned invisibly.
}
\description{
This function provides a summary of a Convolutional Neural Network (CNN) model created using the \code{\link{cnn}} function. It currently replicates the output of the \code{\link{print.citocnn}} method.
}
