% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{summary.citodnnBootstrap}
\alias{summary.citodnnBootstrap}
\title{Summarize Neural Network of class citodnnBootstrap}
\usage{
\method{summary}{citodnnBootstrap}(object, n_permute = NULL, device = NULL, ...)
}
\arguments{
\item{object}{a model of class citodnn created by \code{\link{dnn}} with bootstrapping}

\item{n_permute}{number of permutations performed. Default is \eqn{3 * \sqrt{n}}, where n euqals then number of samples in the training set}

\item{device}{for calculating variable importance and conditional effects}

\item{...}{additional arguments}
}
\value{
summary.citodnnBootstrap returns an object of class "summary.citodnn", a list with components
}
\description{
Performs a Feature Importance calculation based on Permutations
}
\details{
Performs the feature importance calculation as suggested by  Fisher, Rudin, and Dominici (2018), and the mean and standard deviation of the average conditional Effects as suggested by Pichler & Hartig (2023).

Feature importances are in their interpretation similar to a ANOVA. Main and interaction effects are absorbed into the features. Also, feature importances are prone to collinearity between features, i.e. if two features are collinear, the importances might be overestimated.

Average conditional effects (ACE) are similar to marginal effects and approximate linear effects, i.e. their interpretation is similar to effects in a linear regression model.

The standard deviation of the ACE informs about the non-linearity of the feature effects. Higher values correlate with stronger non-linearities.
}
