% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformulate_glmmTMB.R
\name{sumTerms}
\alias{sumTerms}
\title{combine a list of formula terms as a sum}
\usage{
sumTerms(termList)
}
\arguments{
\item{termList}{a list of formula terms}
}
\description{
combine a list of formula terms as a sum
}
