% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\docType{data}
\name{tree_data}
\alias{tree_data}
\title{Tree Data}
\format{
\subsection{\code{tree_data}}{

List with the following objects (500 observations)
\describe{
\item{images}{Satellite images}
\item{Env}{data.frame, Environment (tabulardata)}
\item{Species}{Species 0 or Species 1}
\item{Trees}{Number of Trees}
\item{dbh}{DBH in cm}
}
}
}
\usage{
tree_data
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
