% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{analyze_training}
\alias{analyze_training}
\title{Visualize training of Neural Network afterwards to decide on best performing model.
Creates a plotly figure which allows to zoom in and out on training graph}
\usage{
analyze_training(object)
}
\arguments{
\item{object}{a model created by \code{\link{dnn}}}
}
\value{
a plotly figure
}
\description{
Visualize training of Neural Network afterwards to decide on best performing model.
Creates a plotly figure which allows to zoom in and out on training graph
}
\examples{
\dontrun{
library(cito)
set.seed(222)
validation_set<- sample(c(1:nrow(datasets::iris)),25)

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris[-validation_set,],validation = 0.1)

# show zoomable plot of training and validation losses
analyze_training(nn.fit)

# set model which is used for predictions to model from epoch 46
nn.fit$use_model_epoch <- 46

# Use model on validation set
predictions <- predict(nn.fit, iris[validation_set,])

# Scatterplot
plot(iris[validation_set,]$Sepal.Length,predictions)
}
}
