% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{avgPool}
\alias{avgPool}
\title{Average pooling layer}
\usage{
avgPool(kernel_size = NULL, stride = NULL, padding = NULL)
}
\arguments{
\item{kernel_size}{(int or tuple) size of the kernel in this layer. Use a tuple if the kernel size isn't equal in all dimensions}

\item{stride}{(int or tuple) stride of the kernel in this layer. NULL sets the stride equal to the kernel size. Use a tuple if the stride isn't equal in all dimensions}

\item{padding}{(int or tuple) zero-padding added to both sides of the input. Use a tuple if the padding isn't equal in all dimensions}
}
\value{
S3 object of class \code{"avgPool" "citolayer"}
}
\description{
creates a 'avgPool' 'citolayer' object that is used by \code{\link{create_architecture}}.
}
\details{
This function creates a 'avgPool' 'citolayer' object that is passed to the \code{\link{create_architecture}} function.
The parameters that aren't assigned here (and are therefore still NULL) are filled with the default values passed to \code{\link{create_architecture}}.
}
\examples{
\donttest{
  if(torch::torch_is_installed()){
    library(cito)

    # A average pooling layer where all available parameters are assigned
    # No value will be overwritten by 'create_architecture()'
    layer1 <- avgPool(3, 1, 0)

    # A average pooling layer where only the kernel size is assigned
    # stride and padding are filled with the defaults
    # passed to the 'create_architecture()' function
    layer2 <- avgPool(kernel_size=4)
  }
}
}
\seealso{
\code{\link{create_architecture}}
}
\author{
Armin Schenk
}
