% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{coef.citocnn}
\alias{coef.citocnn}
\title{Returns list of parameters the neural network model currently has in use}
\usage{
\method{coef}{citocnn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{cnn}}}

\item{...}{nothing implemented yet}
}
\value{
list of weights of neural network
}
\description{
Returns list of parameters the neural network model currently has in use
}
\examples{
\donttest{
if(torch::torch_is_installed()){
library(cito)

set.seed(222)

## Data
shapes <- cito:::simulate_shapes(320, 28)
X <- shapes$data
Y <- shapes$labels

## Architecture
architecture <- create_architecture(conv(5), maxPool(), conv(5), maxPool(), linear(10))

## Build and train network
cnn.fit <- cnn(X, Y, architecture, loss = "softmax", epochs = 50, validation = 0.1, lr = 0.05)

# Weights of neural network
coef(cnn.fit)
}
}
}
