% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACE.R
\name{conditionalEffects}
\alias{conditionalEffects}
\title{Calculate average conditional effects}
\usage{
conditionalEffects(
  object,
  interactions = FALSE,
  epsilon = 0.1,
  device = c("cpu", "cuda", "mps"),
  indices = NULL,
  data = NULL,
  type = "response",
  ...
)
}
\arguments{
\item{object}{object of class nn.fit}

\item{interactions}{calculate interactions or not (computationally expensive)}

\item{epsilon}{difference used to calculate derivatives}

\item{device}{which device}

\item{indices}{calculation of effects until which column index}

\item{data}{data which is used to calculate the ACE}

\item{type}{ACE on which scale (response or link)}

\item{...}{additional arguments that are passed to the predict function}
}
\value{
an S3 object of class \code{"conditionalEffects"} is returned.
The list consists of the following attributes:
\item{result}{3-dimensional array with the raw results}
\item{mean}{Matrix, average conditional effects}
\item{abs}{Matrix, summed absolute conditional effects}
\item{sd}{Matrix, standard deviation of the conditional effects}
}
\description{
Calculate average conditional effects
}
\author{
Maximilian Pichler
}
