% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{config_tuning}
\alias{config_tuning}
\title{Config hyperparameter tuning}
\usage{
config_tuning(CV = 5, steps = 10, parallel = FALSE, NGPU = 1)
}
\arguments{
\item{CV}{numeric, specifies k-folded cross validation}

\item{steps}{numeric, number of random tuning steps}

\item{parallel}{numeric, number of parallel cores (tuning steps are parallelized)}

\item{NGPU}{numeric, set if more than one GPU is available, tuning will be parallelized over CPU cores and GPUs, only works for NCPU > 1}
}
\description{
Config hyperparameter tuning
}
