% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_training.R
\name{continue_training}
\alias{continue_training}
\title{Contiinues training of a model for additional periods}
\usage{
continue_training(
  model,
  epochs = 32,
  continue_from = NULL,
  data = NULL,
  device = "cuda",
  changed_params = NULL
)
}
\arguments{
\item{model}{a model created by \code{\link{dnn}}}

\item{epochs}{additional epochs the training should continue for}

\item{continue_from}{define which epoch should be used as starting point for training, 0 if last epoch should be used}

\item{data}{matrix or data.frame if not provided data from original training will be used}

\item{device}{device on which network should be trained on, either "cpu" or "cuda"}

\item{changed_params}{list of arguments to change compared to original training setup, see \code{\link{dnn}} which parameter can be changed}
}
\value{
a model of class cito.dnn same as created by  \code{\link{dnn}}
}
\description{
Contiinues training of a model for additional periods
}
\examples{
/inst/examples/continue_training-example.R

}
