% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_training.R
\name{continue_training}
\alias{continue_training}
\title{Contiinues training of a model for additional periods}
\usage{
continue_training(
  model,
  epochs = 32,
  early_stopping = FALSE,
  lr_scheduler = FALSE
)
}
\arguments{
\item{model}{a model created by \code{\link{dnn}}}

\item{epochs}{additional epochs the training should continue for}

\item{early_stopping}{training stops if validation error n epochs before was lower than in current epoch}

\item{lr_scheduler}{lr scheduler}

\item{...}{additional arguments}
}
\value{
prediction matrix
}
\description{
Contiinues training of a model for additional periods
}
