% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{e}
\alias{e}
\title{Embeddings}
\usage{
e(dim = 1L, weights = NULL, train = TRUE)
}
\arguments{
\item{dim}{integer, embedding dimension}

\item{weights}{matrix, to use custom embedding matrices}

\item{train}{logical, should the embeddings be trained or not}
}
\description{
Can be used for categorical variables, it is more efficient than one-hot encoding
}
