% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PDP}
\alias{PDP}
\title{Partial Dependence Plot (PDP)}
\usage{
PDP(model, variable)
}
\arguments{
\item{model}{a model created by \code{\link{dnn}}}

\item{variable}{variable as string for which the PDP should be done}
}
\description{
Calculates the Partial Dependency Plot for one feature, either numeric or categorical. Returns it as a plot.
}
\details{
Performs the estimation of the partial function \eqn{\hat{f}_S}{}

\eqn{\hat{f}_S(x_S)=\frac{1}{n}\sum_{i=1}^n\hat{f}(x_S,x^{(i)}_{C})}{}

with a Monte Carlo Estimation:

\eqn{\hat{f}_S(x_S)=\frac{1}{n}\sum_{i=1}^n\hat{f}(x_S,x^{(i)}_{C})}{}

If a categorical feature is analyzed, all data instances are used and set to each level.
Then an average is calculated per category and put out in a bar plot.
}
\examples{
\dontrun{
library(cito)

# Build and train  Network
nn.fit<- dnn(Sepal.Length~., data = datasets::iris)

PDP(nn.fit, variable = "Petal.Length")
}
}
\seealso{
\code{\link{ALE}}
}
