% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{plot.citocnn}
\alias{plot.citocnn}
\title{Plot the CNN architecture}
\usage{
\method{plot}{citocnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
Plot the CNN architecture
}
\examples{
\donttest{
if(torch::torch_is_installed()){
library(cito)

set.seed(222)

## Data
shapes <- cito:::simulate_shapes(320, 28)
X <- shapes$data
Y <- shapes$labels

## Architecture
architecture <- create_architecture(conv(5), maxPool(), conv(5), maxPool(), linear(10))

## Build and train network
cnn.fit <- cnn(X, Y, architecture, loss = "softmax", epochs = 50, validation = 0.1, lr = 0.05)

## Structure of Neural Network
plot(cnn.fit)
}
}
}
