% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{plot.citocnn}
\alias{plot.citocnn}
\title{Plot the CNN architecture}
\usage{
\method{plot}{citocnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
Plot the CNN architecture
}
