% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{plot.citodnn}
\alias{plot.citodnn}
\title{Creates graph plot which gives an overview of the network architecture.}
\usage{
\method{plot}{citodnn}(x, node_size = 1, scale_edges = FALSE, ...)
}
\arguments{
\item{x}{a model created by \code{\link{dnn}}}

\item{node_size}{size of node in plot}

\item{scale_edges}{edge weight gets scaled according to other weights (layer specific)}

\item{...}{no further functionality implemented yet}
}
\description{
Creates graph plot which gives an overview of the network architecture.
}
\examples{
/inst/examples/plot.citodnn-example.R
}
