% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmn.R
\name{predict.citommn}
\alias{predict.citommn}
\title{Predict from a fitted mmn model}
\usage{
\method{predict}{citommn}(
  object,
  newdata = NULL,
  type = c("link", "response", "class"),
  device = c("cpu", "cuda", "mps"),
  ...
)
}
\arguments{
\item{object}{a model created by \code{\link{mmn}}}

\item{newdata}{new data for predictions}

\item{type}{which value should be calculated, either raw response, output of link function or predicted class (in case of classification)}

\item{device}{device on which network should be trained on.}

\item{...}{additional arguments}
}
\value{
prediction matrix
}
\description{
Predict from a fitted mmn model
}
