% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.citoarchitecture}
\alias{print.citoarchitecture}
\title{Print class citoarchitecture}
\usage{
\method{print}{citoarchitecture}(x, input_shape, output_shape, ...)
}
\arguments{
\item{x}{an object created by \code{\link{create_architecture}}}

\item{input_shape}{a vector with the dimensions of a single sample (e.g. c(3,28,28))}

\item{output_shape}{the number of nodes in the output layer}

\item{...}{additional arguments}
}
\value{
original object
}
\description{
Print class citoarchitecture
}
