% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.citocnn}
\alias{print.citocnn}
\title{Print class citocnn}
\usage{
\method{print}{citocnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
Print class citocnn
}
