% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{print.citocnn}
\alias{print.citocnn}
\title{Print class citocnn}
\usage{
\method{print}{citocnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
Print class citocnn
}
\examples{
\donttest{
  if(torch::torch_is_installed()){
    library(cito)

    set.seed(222)

    ## Data
    shapes <- simulate_shapes(1000, 100)
    X <- shapes$data
    Y <- shapes$labels

    ## Architecture
    architecture <- create_architecture(conv(4), conv(8), maxPool(), linear(10))

    ## Build and train network
    cnn.fit <- cnn(X, Y, architecture, loss = "softmax", epochs = 50, validation = 0.2)

    # Structure of Neural Network
    print(cnn.fit)
  }
}
}
