% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{print.citodnn}
\alias{print.citodnn}
\title{Print class citodnn}
\usage{
\method{print}{citodnn}(x, ...)
}
\arguments{
\item{x}{a model created by \code{\link{dnn}}}

\item{...}{additional arguments}
}
\value{
prediction matrix
}
\description{
Print class citodnn
}
