% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmn.R
\name{print.citommn}
\alias{print.citommn}
\title{Print a fitted MMN model}
\usage{
\method{print}{citommn}(x, ...)
}
\arguments{
\item{x}{A model created by \code{\link{mmn}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The original model object \code{x}, returned invisibly.
}
\description{
This function prints the architecture of a Multi-Modal Neural Network (MMN) model created using the \code{\link{mmn}} function.
}
