% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACE.R
\name{print.conditionalEffects}
\alias{print.conditionalEffects}
\title{Print average conditional effects}
\usage{
\method{print}{conditionalEffects}(x, ...)
}
\arguments{
\item{x}{print ACE calculated by \code{\link{conditionalEffects}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
Matrix with average conditional effects
}
\description{
Print average conditional effects
}
