% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_shapes}
\alias{simulate_shapes}
\title{Data Simulation for CNN}
\usage{
simulate_shapes(n, size, channels = 1)
}
\arguments{
\item{n}{number of images}

\item{size}{size of the (quadratic) images}

\item{channels}{number of channels the generated data has (in each channel a new rectangle/ellipsoid is created)}
}
\value{
array of dimension (n, 1, size, size)
}
\description{
generates images of rectangles and ellipsoids
}
\details{
This function generates simple data to demonstrate the usage of cnn().
The generated images are of centered rectangles and ellipsoids with random widths and heights.
}
\author{
Armin Schenk
}
