% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{summary.citocnn}
\alias{summary.citocnn}
\title{Summary citocnn}
\usage{
\method{summary}{citocnn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{cnn}}}

\item{...}{additional arguments}
}
\value{
original object x
}
\description{
currently the same as the print.citocnn method.
}
