% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{summary.citodnn}
\alias{summary.citodnn}
\title{Summarize Neural Network of class citodnn}
\usage{
\method{summary}{citodnn}(object, n_permute = 256, ...)
}
\arguments{
\item{object}{a model of class citodnn created by \code{\link{dnn}}}

\item{...}{additional arguments}
}
\value{
summary.glm returns an object of class "summary.citodnn", a list with components
}
\description{
Summarize Neural Network of class citodnn
}
