% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmn.R
\name{summary.citommn}
\alias{summary.citommn}
\title{Summary citommn}
\usage{
\method{summary}{citommn}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{mmn}}}

\item{...}{additional arguments}
}
\value{
original object
}
\description{
currently the same as the print.citommn method.
}
