% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{tune}
\alias{tune}
\title{Tune hyperparameter}
\usage{
tune(lower = NULL, upper = NULL, fixed = NULL, additional = NULL)
}
\arguments{
\item{lower}{numeric, numeric vector, character, lower boundaries of tuning space}

\item{upper}{numeric, numeric vector, character, upper foundaries of tuning space}

\item{fixed}{charachter, used for multi-dimensional hyperparameters such as hidden, which dimensions should be fixed}

\item{additional}{numeric, additional control parameter which sets the value of the fixed argument}
}
\description{
Control hyperparameter tuning
}
